Add swapbuffers and vsync_wait options
Some drivers will automatically prevent tearing when swapbuffers or
copysubbuffer is called, and will throttle drawing at that time as well
(usually to the refresh rate); the --no-wait option can be used in that
case.

And for testing drivers, the --always-swap option can be nice, since it
will force full screen updates, allowing page flipping paths to be
taken, for example.

Signed-off-by: Jesse Barnes <jbarnes@virtuousgeek.org>